refactor:frontend-directories (#15)
* refactor:frontend-directories

* refactor: use path alias
