+++ docs/TODO.MD
@@ -7,6 +7,7 @@
 - [x] feat: filter by categories on gragh
 - [x] refactor: separate component
 - [x] feat: filter link skill count on gragh
+- [x] refactor: frondend directory
 - [] feat: reverse search
 - [] feat: sort categories
 - [] feat: batch damage
