+++ frontend/src/features/skillChaining/components/visualization/SkillFlowChart.css
