+++ frontend/src/features/skillChaining/components/visualization/SmartSearch.tsx
