+++ frontend/src/features/skillChaining/components/visualization/VisualizationContent.tsx
