+++ frontend/src/features/skillChaining/components/visualization/VisualizationInfo.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { getCategoryColor } from './categoryColors';
+import { getCategoryColor } from '../../utils/categoryColors';
 
 interface VisualizationInfoProps {
   categories: string[];
