+++ frontend/src/features/skillChaining/utils/categoryColors.ts
