+++ frontend/vite.config.ts
@@ -6,5 +6,15 @@ export default defineConfig({
   server: {
     port: 5173,
     strictPort: false,
+  },
+  resolve: {
+    alias: {
+      '@': '/src',
+      '@api': '/src/api',
+      '@components': '/src/components',
+      '@features': '/src/features',
+      '@layouts': '/src/layouts',
+      '@utils': '/src/utils'
+    }
   }
 });
